Let's Pivot
Let's make a platform for recording and analyzing market data,
and there could be a companion mobile app that sends notifications
when our system notices potential buying opportunities.

If the quality of the notifications turns out to be good enough in
our own testing, I would even charge for the mobile app.

This is something I would like to have for myself anyway.